# Specialize casting for 86