--- a/timemachine/cpp/src/kernels/k_fixed_point.cuh
+++ b/timemachine/cpp/src/kernels/k_fixed_point.cuh
@@ -11,6 +11,7 @@ RealType __device__ __forceinline__ FIXED_TO_FLOAT_DU_DP(unsigned long long v) {
 // (ytz): courtesy of @scottlegrand/NVIDIA, even faster conversion
 // This was original a hack to improve perf on Maxwell, that is now needed for Ampere
 long long __device__ __forceinline__ real_to_int64(float x) {
+#if __CUDA_ARCH__ == 860
   float z = x * (float)0x1.00000p-32;
   int hi = __float2int_rz( z );                         // First convert high bits
   float delta = x - ((float)0x1.00000p32*((float)hi));  // Check remainder sign
@@ -20,6 +21,9 @@ long long __device__ __forceinline__ real_to_int64(float x) {
   hi -= test;                                           // Two's complement correction
   long long res = __double_as_longlong(__hiloint2double(hi,lo)); // Return 64-bit result
   return res;
+#else
+    return static_cast<long long>(x);
+#endif
 }
 
 // (ytz): reference version, left here for pedagogical reasons, do not remove.