@proteneer 2021-03-08
@dmclark17 Do you know if llprintf and static_cast<long long> are identical?