@proteneer 2021-03-08
@dmclark17 Sorry that was a typo, I meant `llrintf` that was originally documented in slg's code. The C API is here:

https://linux.die.net/man/3/llrintf