@dmclark17 2021-03-08
Okay! Sorry I didn't put that together. The [CUDA documentation](https://docs.nvidia.com/cuda/cuda-math-api/group__CUDA__MATH__SINGLE.html#group__CUDA__MATH__SINGLE_1g7d4af230b5deee73fbfa9801f44f0616) for `llrintf ` says it will round to the nearest integer value. I believe the static cast should truncate the number.