@proteneer 2021-03-08
Thanks for the clarification. So it sounds like llrintf has less round off error than static_cast, but since net forces are conservative (we have roughly an equal number of +x and -x, and the net force is zero), a static_cast/truncation is probably okay. Maybe there's a way to optimize this even further then...