@proteneer 2021-03-08
weird - `llrintf` is actually faster than `static_cast`:

```
dhfr-apo: N=23558 speed: 320.12ns/day dt: 1.5fs (ran 100000 steps in 40.49s)
building a protein system with 1758 protein atoms and 7047 water atoms
hif2a-apo: N=8805 speed: 570.87ns/day dt: 1.5fs (ran 100000 steps in 22.71s)
WARNING:absl:No GPU/TPU found, falling back to CPU. (Set TF_CPP_MIN_LOG_LEVEL=0 and rerun for more info.)
hif2a-rbfe-with-du-dp: N=8840 speed: 489.62ns/day dt: 1.5fs (ran 100000 steps in 26.47s)
hif2a-rbfe-du-dl-freq-0: N=8840 speed: 491.30ns/day dt: 1.5fs (ran 100000 steps in 26.38s)
hif2a-rbfe-du-dl-freq-1: N=8840 speed: 465.31ns/day dt: 1.5fs (ran 100000 steps in 27.86s)
hif2a-rbfe-du-dl-freq-5: N=8840 speed: 485.25ns/day dt: 1.5fs (ran 100000 steps in 26.71s)
solvent-apo: N=6282 speed: 715.23ns/day dt: 1.5fs (ran 100000 steps in 18.12s)
solvent-rbfe-with-du-dp: N=6317 speed: 598.54ns/day dt: 1.5fs (ran 100000 steps in 21.66s)
solvent-rbfe-du-dl-freq-0: N=6317 speed: 601.11ns/day dt: 1.5fs (ran 100000 steps in 21.57s)
solvent-rbfe-du-dl-freq-1: N=6317 speed: 557.02ns/day dt: 1.5fs (ran 100000 steps in 23.27s)
solvent-rbfe-du-dl-freq-5: N=6317 speed: 590.52ns/day dt: 1.5fs (ran 100000 steps in 21.95s)
```