@proteneer 2021-03-09
@maxentile pretty small change, main change is to consistently switch to round-to-nearest-even mode as opposed to round-towards-zero (truncation) when going from float32 to int64s. Going from int64 to float32 is unchanged.